@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <SignAssembly>false</SignAssembly>
     <LangVersion>7.3</LangVersion>
   </PropertyGroup>
@@ -14,11 +14,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="GDAPI" Version="1.0.0" />
+    <PackageReference Include="GDAPI" Version="1.1.1"/>
     <PackageReference Include="ppy.squirrel.windows" Version="1.9.0.4" />
-    <PackageReference Include="ppy.osu.Framework" Version="2019.711.0" />
+    <PackageReference Include="ppy.osu.Framework" Version="2019.1029.0"/>
     <PackageReference Include="SharpRaven" Version="2.4.1-unstable0015" />
-    <PackageReference Include="SixLabors.Core" Version="1.0.0-beta0007" />
+    <PackageReference Include="SixLabors.Core" Version="1.0.0-beta0008"/>
   </ItemGroup>
 
   <ItemGroup>
