@@ -1,34 +1,27 @@
-﻿using GDE.App.Main.Containers.KeyBindingContainers;
+﻿using System.Collections.Generic;
+using GDAPI.Application;
+using GDAPI.Application.Editors;
+using GDAPI.Objects.GeometryDash.General;
+using GDE.App.Main.Containers.KeyBindingContainers;
 using GDE.App.Main.Objects;
 using GDE.App.Main.Screens.Edit;
 using GDE.App.Main.Screens.Edit.Components;
 using GDE.App.Main.UI;
-using GDAPI.Application;
-using GDAPI.Application.Editor;
-using GDAPI.Utilities.Objects.GeometryDash;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Input.Bindings;
-using System.Collections.Generic;
 
 namespace GDE.App.Main.Levels
 {
     public class LevelPreview : Container<ObjectBase>, IKeyBindingHandler<GlobalAction>, IDraggable
     {
-        private EditorScreen editorScreen;
-
         private readonly int i;
 
         private Database database;
+        private EditorScreen editorScreen;
 
         private bool modifier;
-
-        public IReadOnlyList<ObjectBase> Objects => Children;
-
-        public bool Draggable => true;
-
-        public Level Level => database.UserLevels[i];
 
         public LevelPreview(EditorScreen editorScreen, int index)
         {
@@ -38,21 +31,17 @@ namespace GDE.App.Main.Levels
             AutoSizeAxes = Axes.Both;
         }
 
-        [BackgroundDependencyLoader]
-        private void load(DatabaseCollection databases)
-        {
-            database = databases[0];
+        public IReadOnlyList<ObjectBase> Objects => Children;
 
-            foreach (var o in Level.LevelObjects)
-                Add(new ObjectBase(o));
-        }
+        public Level Level => database.UserLevels[i];
+
+        public bool Draggable => true;
 
         public bool OnPressed(GlobalAction action)
         {
             var val = modifier ? Editor.SmallMovementStep : Editor.NormalMovementStep;
 
             foreach (var i in Objects)
-            {
                 if (i.State == SelectionState.Selected)
                     switch (action)
                     {
@@ -69,7 +58,6 @@ namespace GDE.App.Main.Levels
                             i.ObjectY -= val;
                             break;
                     }
-            }
 
             switch (action)
             {
@@ -81,6 +69,18 @@ namespace GDE.App.Main.Levels
             return true;
         }
 
-        public bool OnReleased(GlobalAction action) => true;
+        public bool OnReleased(GlobalAction action)
+        {
+            return true;
+        }
+
+        [BackgroundDependencyLoader]
+        private void load(DatabaseCollection databases)
+        {
+            database = databases[0];
+
+            foreach (var o in Level.LevelObjects)
+                Add(new ObjectBase(o));
+        }
     }
 }
