@@ -1,6 +1,8 @@
-﻿using GDE.App.Main.UI;
-using GDAPI.Application.Editor;
-using GDAPI.Utilities.Objects.GeometryDash.LevelObjects;
+﻿using System;
+using System.Collections.Generic;
+using GDAPI.Application.Editors;
+using GDAPI.Objects.GeometryDash.LevelObjects;
+using GDE.App.Main.UI;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Sprites;
@@ -9,90 +11,28 @@ using osu.Framework.Input.Events;
 using osu.Framework.Logging;
 using osuTK;
 using osuTK.Graphics;
-using System;
-using System.Collections.Generic;
 
 namespace GDE.App.Main.Objects
 {
-    ///<summary>A drawable <seealso cref="GeneralObject"/>.</summary>
+    /// <summary>A drawable <seealso cref="GeneralObject" />.</summary>
     public class ObjectBase : Sprite
     {
-        private TextureStore textureStore;
-        [Resolved]
-        private Editor editor { get; set; }
+        public static List<ObjectBase> DrawableSelectedObjects;
+        public readonly ObjectBase Object;
 
         public GeneralObject LevelObject;
-        public readonly ObjectBase Object;
-        public SelectionState State;
-        public EventHandler Selected;
         public bool Selectable = true;
-        public static List<ObjectBase> DrawableSelectedObjects;
+        public EventHandler Selected;
+        public SelectionState State;
+        private TextureStore textureStore;
 
         static ObjectBase()
         {
             DrawableSelectedObjects = new List<ObjectBase>();
         }
 
-        #region Level Object Variables
-        ///<summary>The ID of the object.</summary>
-        public int ObjectID
-        {
-            get => LevelObject.ObjectID;
-            set => UpdateObjectID(LevelObject.ObjectID = value);
-        }
-        ///<summary>The X position of the object.</summary>
-        public double ObjectX
-        {
-            get => LevelObject.X;
-            set => UpdateObjectX(LevelObject.X = value);
-        }
-        ///<summary>The Y position of the object.</summary>
-        public double ObjectY
-        {
-            get => LevelObject.Y;
-            set => UpdateObjectY(LevelObject.Y = value);
-        }
-        ///<summary>Represents whether the object is flipped horizontally or not.</summary>
-        public bool FlippedHorizontally
-        {
-            get => LevelObject.FlippedHorizontally;
-            set => UpdateFlippedHorizontally(LevelObject.FlippedHorizontally = value);
-        }
-        ///<summary>Represents whether the object is flipped vertically or not.</summary>
-        public bool FlippedVertically
-        {
-            get => LevelObject.FlippedVertically;
-            set => UpdateFlippedVertically(LevelObject.FlippedVertically = value);
-        }
-        ///<summary>The rotation of the object.</summary>
-        public double ObjectRotation
-        {
-            get => LevelObject.Rotation;
-            set => UpdateObjectRotation(LevelObject.Rotation = value);
-        }
-        ///<summary>The scaling of the object.</summary>
-        public double ObjectScaling
-        {
-            get => LevelObject.Scaling;
-            set => UpdateObjectScaling(LevelObject.Scaling = value);
-        }
-        ///<summary>The Editor Layer 1 of the object.</summary>
-        public int EL1
-        {
-            get => LevelObject.EL1;
-            set => LevelObject.EL1 = value;
-        }
-        ///<summary>The Editor Layer 2 of the object.</summary>
-        public int EL2
-        {
-            get => LevelObject.EL2;
-            set => LevelObject.EL2 = value;
-        }
-        #endregion
-
-        /// <summary>Initializes a new instance of the <seealso cref="ObjectBase"/> class.</summary>
+        /// <summary>Initializes a new instance of the <seealso cref="ObjectBase" /> class.</summary>
         public ObjectBase(GeneralObject o)
-            : base()
         {
             Anchor = Anchor.Centre;
             Origin = Anchor.Centre;
@@ -100,8 +40,13 @@ namespace GDE.App.Main.Objects
             Size = new Vector2(30);
             UpdateObject(LevelObject = o);
         }
-        /// <summary>Initializes a new instance of the <seealso cref="ObjectBase"/> class.</summary>
-        public ObjectBase(int ID) : this(GeneralObject.GetNewObjectInstance(ID)) { }
+
+        /// <summary>Initializes a new instance of the <seealso cref="ObjectBase" /> class.</summary>
+        public ObjectBase(int ID) : this(GeneralObject.GetNewObjectInstance(ID))
+        {
+        }
+
+        [Resolved] private Editor editor { get; set; }
 
         [BackgroundDependencyLoader]
         private void load(TextureStore ts)
@@ -121,16 +66,44 @@ namespace GDE.App.Main.Objects
             UpdateObjectScaling(o.Scaling);
         }
 
-        private void UpdateObjectID(int value) => Texture = value > 0 ? textureStore?.Get($"Objects/{value}.png") : null;
-        private void UpdateObjectX(double value) => X = (float)value;
-        private void UpdateObjectY(double value) => Y = -(float)value;
-        private void UpdateFlippedHorizontally(bool value) => Width = SetSign(Width, !value);
-        private void UpdateFlippedVertically(bool value) => Height = SetSign(Height, !value);
-        private void UpdateObjectRotation(double value) => Rotation = (float)value;
-        private void UpdateObjectScaling(double value) => Scale = new Vector2((float)value);
+        private void UpdateObjectID(int value)
+        {
+            Texture = value > 0 ? textureStore?.Get($"Objects/{value}.png") : null;
+        }
+
+        private void UpdateObjectX(double value)
+        {
+            X = (float) value;
+        }
+
+        private void UpdateObjectY(double value)
+        {
+            Y = -(float) value;
+        }
+
+        private void UpdateFlippedHorizontally(bool value)
+        {
+            Width = SetSign(Width, !value);
+        }
+
+        private void UpdateFlippedVertically(bool value)
+        {
+            Height = SetSign(Height, !value);
+        }
+
+        private void UpdateObjectRotation(double value)
+        {
+            Rotation = (float) value;
+        }
+
+        private void UpdateObjectScaling(double value)
+        {
+            Scale = new Vector2((float) value);
+        }
+
         private float SetSign(float value, bool sign)
         {
-            if (!sign ^ value < 0)
+            if (!sign ^ (value < 0))
                 return -value;
             return value;
         }
@@ -161,5 +134,72 @@ namespace GDE.App.Main.Objects
 
             return base.OnClick(e);
         }
+
+        #region Level Object Variables
+
+        ///<summary>The ID of the object.</summary>
+        public int ObjectID
+        {
+            get => LevelObject.ObjectID;
+            set => UpdateObjectID(LevelObject.ObjectID = value);
+        }
+
+        ///<summary>The X position of the object.</summary>
+        public double ObjectX
+        {
+            get => LevelObject.X;
+            set => UpdateObjectX(LevelObject.X = value);
+        }
+
+        ///<summary>The Y position of the object.</summary>
+        public double ObjectY
+        {
+            get => LevelObject.Y;
+            set => UpdateObjectY(LevelObject.Y = value);
+        }
+
+        ///<summary>Represents whether the object is flipped horizontally or not.</summary>
+        public bool FlippedHorizontally
+        {
+            get => LevelObject.FlippedHorizontally;
+            set => UpdateFlippedHorizontally(LevelObject.FlippedHorizontally = value);
+        }
+
+        ///<summary>Represents whether the object is flipped vertically or not.</summary>
+        public bool FlippedVertically
+        {
+            get => LevelObject.FlippedVertically;
+            set => UpdateFlippedVertically(LevelObject.FlippedVertically = value);
+        }
+
+        ///<summary>The rotation of the object.</summary>
+        public double ObjectRotation
+        {
+            get => LevelObject.Rotation;
+            set => UpdateObjectRotation(LevelObject.Rotation = value);
+        }
+
+        ///<summary>The scaling of the object.</summary>
+        public double ObjectScaling
+        {
+            get => LevelObject.Scaling;
+            set => UpdateObjectScaling(LevelObject.Scaling = value);
+        }
+
+        ///<summary>The Editor Layer 1 of the object.</summary>
+        public int EL1
+        {
+            get => LevelObject.EL1;
+            set => LevelObject.EL1 = value;
+        }
+
+        ///<summary>The Editor Layer 2 of the object.</summary>
+        public int EL2
+        {
+            get => LevelObject.EL2;
+            set => LevelObject.EL2 = value;
+        }
+
+        #endregion
     }
 }
