@@ -1,8 +1,8 @@
-﻿using GDAPI.Utilities.Enumerations;
+﻿using System;
+using GDAPI.Enumerations;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osuTK;
-using System;
 
 namespace GDE.App.Main.Screens.Edit.Components.IDMigration
 {
@@ -10,25 +10,23 @@ namespace GDE.App.Main.Screens.Edit.Components.IDMigration
     {
         public const int DefaultHeight = 32;
 
-        private IDMigrationTabItem[] tabItems = new IDMigrationTabItem[4];
+        private readonly IDMigrationTabItem[] tabItems = new IDMigrationTabItem[4];
         private IDMigrationTabItem currentTab;
 
         private FillFlowContainer itemsContainer;
 
-        public event Action<IDMigrationMode> TabSelected;
-
         public IDMigrationTabControl()
-            : base()
         {
-            for (int i = 0; i < 4; i++)
+            for (var i = 0; i < 4; i++)
             {
-                tabItems[i] = new IDMigrationTabItem((IDMigrationMode)i)
+                tabItems[i] = new IDMigrationTabItem((IDMigrationMode) i)
                 {
                     // WHY THE FUCK DOES THIS NOT WORK?
                     Y = 20
                 };
                 tabItems[i].TabSelected += HandleTabSelected;
             }
+
             SelectIDMigrationMode(IDMigrationMode.Groups);
 
             RelativeSizeAxes = Axes.X;
@@ -43,15 +41,21 @@ namespace GDE.App.Main.Screens.Edit.Components.IDMigration
                     Direction = FillDirection.Horizontal,
                     RelativeSizeAxes = Axes.Both,
                     Spacing = new Vector2(5),
-                    Padding = new MarginPadding { Left = 10 },
-                    Children = tabItems,
+                    Padding = new MarginPadding {Left = 10},
+                    Children = tabItems
                 }
             };
         }
 
-        public IDMigrationTabItem this[IDMigrationMode mode] => tabItems[(int)mode];
+        public IDMigrationTabItem this[IDMigrationMode mode] => tabItems[(int) mode];
 
-        public void SelectIDMigrationMode(IDMigrationMode newMode) => this[newMode].Selected = true;
+        public event Action<IDMigrationMode> TabSelected;
+
+        public void SelectIDMigrationMode(IDMigrationMode newMode)
+        {
+            this[newMode].Selected = true;
+        }
+
         public void HandleTabSelected(IDMigrationMode newMode)
         {
             if (currentTab != null)
@@ -59,6 +63,7 @@ namespace GDE.App.Main.Screens.Edit.Components.IDMigration
                 currentTab.MoveToOffset(new Vector2(0, -10), 500, Easing.OutQuint);
                 currentTab.Selected = false;
             }
+
             (currentTab = this[newMode]).MoveToOffset(new Vector2(0, 10), 500, Easing.InQuint);
             TabSelected?.Invoke(newMode);
         }
