@@ -1,4 +1,8 @@
-﻿using DiscordRPC;
+﻿using System.Collections.Generic;
+using DiscordRPC;
+using GDAPI.Application;
+using GDAPI.Application.Editors;
+using GDAPI.Objects.GeometryDash.General;
 using GDE.App.Main.Colors;
 using GDE.App.Main.Levels;
 using GDE.App.Main.Objects;
@@ -7,9 +11,6 @@ using GDE.App.Main.Screens.Edit.Components.Menu;
 using GDE.App.Main.Tools;
 using GDE.App.Main.UI;
 using GDE.App.Main.UI.FileDialogComponents;
-using GDAPI.Application;
-using GDAPI.Application.Editor;
-using GDAPI.Utilities.Objects.GeometryDash;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
@@ -20,93 +21,27 @@ using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Input.Events;
 using osu.Framework.Screens;
 using osuTK;
-using System.Collections.Generic;
 
 namespace GDE.App.Main.Screens.Edit
 {
     public class EditorScreen : Screen
     {
-        private TextureStore texStore;
-        private Sprite background;
-        private int i;
+        private readonly Sprite background;
+        private readonly Camera camera;
+        public readonly Editor Editor;
+        private readonly int i;
 
-        private Database database;
-        private LevelPreview preview;
-        private Level level => database.UserLevels[i];
-
-        private Grid grid;
-        private Camera camera;
-        private EditorTools tools;
-
-        private IDMigrationPanel IDMigrationPanel;
+        private readonly IDMigrationPanel IDMigrationPanel;
 
         public readonly Bindable<OpenFileDialog> OpenFileDialogBindable = new Bindable<OpenFileDialog>();
+        private readonly LevelPreview preview;
         public readonly Bindable<SaveFileDialog> SaveFileDialogBindable = new Bindable<SaveFileDialog>();
+        private readonly EditorTools tools;
 
-        public readonly Editor Editor;
+        private Database database;
 
-        [BackgroundDependencyLoader]
-        private void load(DatabaseCollection databases, TextureStore ts)
-        {
-            database = databases[0];
-
-            texStore = ts;
-            background.Texture = texStore.Get("Backgrounds/game_bg_01_001-uhd.png");
-
-            Anchor = Anchor.Centre;
-            Origin = Anchor.Centre;
-
-            AddMenuItems();
-        }
-
-        private void AddMenuItems()
-        {
-            EditorMenuBar menuBar;
-
-            var fileMenuItems = new List<MenuItem>
-            {
-                new EditorMenuItem("Save", Save, MenuItemType.Highlighted),
-                new EditorMenuItem("Save & Exit", SaveAndExit, MenuItemType.Standard),
-                new EditorMenuItemSpacer(),
-                new EditorMenuItem("Exit", this.Exit, MenuItemType.Destructive),
-            };
-            var editMenuItems = new List<MenuItem>
-            {
-                // Undo, redo
-            };
-            var macrosMenuItems = new List<MenuItem>
-            {
-                new EditorMenuItem("Migrate IDs", IDMigrationPanel.ToggleVisibility, MenuItemType.Standard),
-            };
-
-            AddInternal(new Container
-            {
-                Name = "Top bar",
-                RelativeSizeAxes = Axes.X,
-                Height = 40,
-                Child = menuBar = new EditorMenuBar
-                {
-                    Anchor = Anchor.CentreLeft,
-                    Origin = Anchor.CentreLeft,
-                    RelativeSizeAxes = Axes.Both,
-                    Items = new[]
-                    {
-                        new MenuItem("File")
-                        {
-                            Items = fileMenuItems
-                        },
-                        new MenuItem("Edit")
-                        {
-                            Items = editMenuItems
-                        },
-                        new MenuItem("Macros")
-                        {
-                            Items = macrosMenuItems
-                        },
-                    }
-                }
-            });
-        }
+        private Grid grid;
+        private TextureStore texStore;
 
         public EditorScreen(int index, Level level)
         {
@@ -150,7 +85,7 @@ namespace GDE.App.Main.Screens.Edit
                         {
                             Anchor = Anchor.Centre,
                             Origin = Anchor.Centre
-                        },
+                        }
                     }
                 },
                 tools = new EditorTools(preview, camera)
@@ -179,8 +114,70 @@ namespace GDE.App.Main.Screens.Edit
                 {
                     Anchor = Anchor.Centre,
                     Origin = Anchor.Centre,
-                    LockDrag = true,
-                },
+                    LockDrag = true
+                }
+            });
+        }
+
+        private Level level => database.UserLevels[i];
+
+        [BackgroundDependencyLoader]
+        private void load(DatabaseCollection databases, TextureStore ts)
+        {
+            database = databases[0];
+
+            texStore = ts;
+            background.Texture = texStore.Get("Backgrounds/game_bg_01_001-uhd.png");
+
+            Anchor = Anchor.Centre;
+            Origin = Anchor.Centre;
+
+            AddMenuItems();
+        }
+
+        private void AddMenuItems()
+        {
+            EditorMenuBar menuBar;
+
+            var fileMenuItems = new List<MenuItem>
+            {
+                new EditorMenuItem("Save", Save, MenuItemType.Highlighted),
+                new EditorMenuItem("Save & Exit", SaveAndExit),
+                new EditorMenuItemSpacer(),
+                new EditorMenuItem("Exit", this.Exit, MenuItemType.Destructive)
+            };
+            var editMenuItems = new List<MenuItem>();
+            var macrosMenuItems = new List<MenuItem>
+            {
+                new EditorMenuItem("Migrate IDs", IDMigrationPanel.ToggleVisibility)
+            };
+
+            AddInternal(new Container
+            {
+                Name = "Top bar",
+                RelativeSizeAxes = Axes.X,
+                Height = 40,
+                Child = menuBar = new EditorMenuBar
+                {
+                    Anchor = Anchor.CentreLeft,
+                    Origin = Anchor.CentreLeft,
+                    RelativeSizeAxes = Axes.Both,
+                    Items = new[]
+                    {
+                        new MenuItem("File")
+                        {
+                            Items = fileMenuItems
+                        },
+                        new MenuItem("Edit")
+                        {
+                            Items = editMenuItems
+                        },
+                        new MenuItem("Macros")
+                        {
+                            Items = macrosMenuItems
+                        }
+                    }
+                }
             });
         }
 
@@ -193,16 +190,13 @@ namespace GDE.App.Main.Screens.Edit
                 preview.Add(new ObjectBase(cloned));
                 return true;
             }
-            else
-                return false;
+
+            return false;
         }
 
         protected override bool OnDrag(DragEvent e)
         {
-            foreach (var child in camera.Children)
-            {
-                child.Position += e.Delta;
-            }
+            foreach (var child in camera.Children) child.Position += e.Delta;
 
             return base.OnDrag(e);
         }
@@ -212,6 +206,10 @@ namespace GDE.App.Main.Screens.Edit
             Save();
             this.Exit();
         }
-        private void Save() => Editor.Save(database, i);
+
+        private void Save()
+        {
+            Editor.Save(database, i);
+        }
     }
 }
