@@ -1,17 +1,17 @@
-﻿using GDE.App.Main.Containers.KeyBindingContainers;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using GDAPI.Enumerations;
+using GDE.App.Main.Containers.KeyBindingContainers;
 using GDE.App.Main.UI.Containers;
-using GDAPI.Utilities.Enumerations;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osuTK;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
 using static GDE.App.Main.Colors.GDEColors;
-using static GDAPI.Utilities.Functions.General.PathExpansionPack;
+using static GDAPI.Functions.General.PathExpansionPack;
 using static System.Math;
 
 namespace GDE.App.Main.UI.FileDialogComponents
@@ -19,50 +19,26 @@ namespace GDE.App.Main.UI.FileDialogComponents
     public class DirectoryItemContainer : KeyBindingActionContainer<FileDialogAction>
     {
         public const float ItemSpacing = 2.5f;
+        public readonly Bindable<string> CurrentDirectoryBindable = new Bindable<string>("");
 
-        private int? currentSelectionIndex;
-        private string currentlyLoadedDirectory;
+        public readonly Bindable<DrawableItem> CurrentSelection = new Bindable<DrawableItem>();
+        private readonly FillFlowContainer fileFillFlowContainer;
 
-        private GDEScrollContainer scrollContainer;
-        private FadeSearchContainer fileContainer;
-        private FillFlowContainer fileFillFlowContainer;
+        private readonly GDEScrollContainer scrollContainer;
 
-        private DrawableItem currentSelection;
+        public readonly Bindable<string> SelectedItemBindable = new Bindable<string>();
 
         /// <summary>The button that performs the file dialog's action.</summary>
         protected GDEButton ActionButton;
 
-        public readonly Bindable<DrawableItem> CurrentSelection = new Bindable<DrawableItem>();
+        private string currentlyLoadedDirectory;
 
-        public readonly Bindable<string> SelectedItemBindable = new Bindable<string>();
-        public readonly Bindable<string> CurrentDirectoryBindable = new Bindable<string>("");
+        private DrawableItem currentSelection;
 
-        public event Action PerformActionRequested;
-
-        public string SelectedItem
-        {
-            get => SelectedItemBindable.Value;
-            set => SelectedItemBindable.Value = value;
-        }
-        public string CurrentDirectory
-        {
-            get => CurrentDirectoryBindable.Value;
-            set => CurrentDirectoryBindable.Value = FixDirectoryPath(value);
-        }
-
-        public DrawableItem CurrentlySelectedItem
-        {
-            get => CurrentSelection.Value;
-            set => CurrentSelection.Value = value;
-        }
-        public int CurrentSelectionIndex
-        {
-            get => currentSelectionIndex ?? (currentSelectionIndex = fileFillFlowContainer.Children.ToList().IndexOf(CurrentlySelectedItem)).Value;
-            set => CurrentlySelectedItem = fileFillFlowContainer.Children[(currentSelectionIndex = value).Value] as DrawableItem;
-        }
+        private int? currentSelectionIndex;
+        private FadeSearchContainer fileContainer;
 
         public DirectoryItemContainer()
-            : base()
         {
             Children = new Drawable[]
             {
@@ -76,7 +52,7 @@ namespace GDE.App.Main.UI.FileDialogComponents
                         new Box
                         {
                             RelativeSizeAxes = Axes.Both,
-                            Colour = FromHex("121212"),
+                            Colour = FromHex("121212")
                         },
                         new Container
                         {
@@ -96,15 +72,15 @@ namespace GDE.App.Main.UI.FileDialogComponents
                                             Direction = FillDirection.Vertical,
                                             Spacing = new Vector2(0, ItemSpacing),
                                             RelativeSizeAxes = Axes.X,
-                                            AutoSizeAxes = Axes.Y,
+                                            AutoSizeAxes = Axes.Y
                                         }
-                                    },
+                                    }
                                 },
-                                ScrollOnKeyDown = false,
-                            },
-                        },
+                                ScrollOnKeyDown = false
+                            }
+                        }
                     }
-                },
+                }
             };
 
             CurrentSelection.ValueChanged += HandleSelectionChanged;
@@ -112,12 +88,65 @@ namespace GDE.App.Main.UI.FileDialogComponents
             SelectedItemBindable.ValueChanged += HandleItemChanged;
         }
 
-        public void NavigateUp() => NavigateTo(CurrentSelectionIndex - 1);
-        public void NavigateDown() => NavigateTo(CurrentSelectionIndex + 1);
-        public void NavigatePageUp() => NavigateTo(CurrentSelectionIndex - GetItemCountPerPage());
-        public void NavigatePageDown() => NavigateTo(CurrentSelectionIndex + GetItemCountPerPage());
-        public void NavigateToStart() => NavigateTo(0);
-        public void NavigateToEnd() => NavigateTo(fileFillFlowContainer.Count - 1);
+        public string SelectedItem
+        {
+            get => SelectedItemBindable.Value;
+            set => SelectedItemBindable.Value = value;
+        }
+
+        public string CurrentDirectory
+        {
+            get => CurrentDirectoryBindable.Value;
+            set => CurrentDirectoryBindable.Value = FixDirectoryPath(value);
+        }
+
+        public DrawableItem CurrentlySelectedItem
+        {
+            get => CurrentSelection.Value;
+            set => CurrentSelection.Value = value;
+        }
+
+        public int CurrentSelectionIndex
+        {
+            get => currentSelectionIndex ??
+                   (currentSelectionIndex = fileFillFlowContainer.Children.ToList().IndexOf(CurrentlySelectedItem))
+                   .Value;
+            set => CurrentlySelectedItem =
+                fileFillFlowContainer.Children[(currentSelectionIndex = value).Value] as DrawableItem;
+        }
+
+        public event Action PerformActionRequested;
+
+        public void NavigateUp()
+        {
+            NavigateTo(CurrentSelectionIndex - 1);
+        }
+
+        public void NavigateDown()
+        {
+            NavigateTo(CurrentSelectionIndex + 1);
+        }
+
+        public void NavigatePageUp()
+        {
+            NavigateTo(CurrentSelectionIndex - GetItemCountPerPage());
+        }
+
+        public void NavigatePageDown()
+        {
+            NavigateTo(CurrentSelectionIndex + GetItemCountPerPage());
+        }
+
+        public void NavigateToStart()
+        {
+            NavigateTo(0);
+        }
+
+        public void NavigateToEnd()
+        {
+            NavigateTo(fileFillFlowContainer.Count - 1);
+        }
+
         public void NavigateToPreviousDirectory()
         {
             var dirs = AnalyzePath(CurrentDirectory);
@@ -126,6 +155,7 @@ namespace GDE.App.Main.UI.FileDialogComponents
 
             CurrentDirectory = ConcatenateDirectoryPath(dirs.SkipLast(2).ToList());
         }
+
         public void NavigateTo(int index)
         {
             if (fileFillFlowContainer.Count > 0)
@@ -140,28 +170,32 @@ namespace GDE.App.Main.UI.FileDialogComponents
                 SelectItem(value.NewValue);
             SelectedItem = value.NewValue?.ItemName;
         }
+
         public void HandleItemChanged(ValueChangedEvent<string> value)
         {
             var newItem = value.NewValue;
 
-            if (CurrentlySelectedItem != null && fileFillFlowContainer.Contains(CurrentlySelectedItem) && CurrentlySelectedItem.ItemName == newItem && CurrentlySelectedItem.Selected)
+            if (CurrentlySelectedItem != null && fileFillFlowContainer.Contains(CurrentlySelectedItem) &&
+                CurrentlySelectedItem.ItemName == newItem && CurrentlySelectedItem.Selected)
                 return;
 
             currentSelectionIndex = null;
 
             if (newItem?.Length > 0)
-            {
                 if (fileFillFlowContainer.Children.ToList().Find(MatchesItem) is DrawableItem item)
                 {
                     SelectItem(item);
                     return;
                 }
-            }
 
             CurrentlySelectedItem = null;
 
-            bool MatchesItem(Drawable item) => (item as DrawableItem)?.ItemName == newItem;
+            bool MatchesItem(Drawable item)
+            {
+                return (item as DrawableItem)?.ItemName == newItem;
+            }
         }
+
         public void HandleDirectoryChanged(ValueChangedEvent<string> value)
         {
             try
@@ -182,7 +216,11 @@ namespace GDE.App.Main.UI.FileDialogComponents
             item.Selected = true;
         }
 
-        private void ForceUpdateItemList() => UpdateItemList(true);
+        private void ForceUpdateItemList()
+        {
+            UpdateItemList(true);
+        }
+
         private void UpdateItemList(bool forceUpdate = false)
         {
             // Mainly to prevent reloading the currently displayed items if the directory is reset on exception
@@ -199,7 +237,8 @@ namespace GDE.App.Main.UI.FileDialogComponents
                 fileFillFlowContainer.Clear();
 
                 foreach (var d in directories)
-                    fileFillFlowContainer.Add(GetNewDrawableItem(GetIndividualItemName(d.Name), PathItemType.Directory));
+                    fileFillFlowContainer.Add(GetNewDrawableItem(GetIndividualItemName(d.Name),
+                        PathItemType.Directory));
                 foreach (var f in files)
                     fileFillFlowContainer.Add(GetNewDrawableItem(GetIndividualItemName(f.Name), PathItemType.File));
             }
@@ -211,7 +250,8 @@ namespace GDE.App.Main.UI.FileDialogComponents
 
                 foreach (var d in drives)
                     if (d.IsReady)
-                        fileFillFlowContainer.Add(GetNewDrawableItem(GetIndividualItemName(d.Name.Remove(d.Name.Length - 1)), PathItemType.Volume));
+                        fileFillFlowContainer.Add(GetNewDrawableItem(
+                            GetIndividualItemName(d.Name.Remove(d.Name.Length - 1)), PathItemType.Volume));
             }
 
             currentlyLoadedDirectory = CurrentDirectory;
@@ -225,39 +265,53 @@ namespace GDE.App.Main.UI.FileDialogComponents
                 CurrentlySelectedItem.Selected = false;
             CurrentlySelectedItem = selectedItem;
         }
+
         private void HandleClick(DrawableItem clickedItem)
         {
             if (!clickedItem.Selected)
                 CurrentlySelectedItem = null;
         }
-        private void HandleDoubleClick(DrawableItem doubleClickedItem) => PerformAction();
 
-        private void PerformAction() => PerformActionRequested?.Invoke();
+        private void HandleDoubleClick(DrawableItem doubleClickedItem)
+        {
+            PerformAction();
+        }
 
-        private int GetItemCountPerPage() => (int)(DrawHeight / (DrawableItem.DefaultHeight + ItemSpacing) + 0.5);
+        private void PerformAction()
+        {
+            PerformActionRequested?.Invoke();
+        }
+
+        private int GetItemCountPerPage()
+        {
+            return (int) (DrawHeight / (DrawableItem.DefaultHeight + ItemSpacing) + 0.5);
+        }
 
         protected override void InitializeActionDictionary()
         {
             // Capacity is greater than the total actions to allow future improvements without *constantly* having to change the constant
             Actions = new Dictionary<FileDialogAction, Action>(20)
             {
-                { FileDialogAction.NavigateUp, NavigateUp },
-                { FileDialogAction.NavigateDown, NavigateDown },
-                { FileDialogAction.NavigatePageUp, NavigatePageUp },
-                { FileDialogAction.NavigatePageDown, NavigatePageDown },
-                { FileDialogAction.NavigateToStart, NavigateToStart },
-                { FileDialogAction.NavigateToEnd, NavigateToEnd },
-                { FileDialogAction.NavigateToPreviousDirectory, NavigateToPreviousDirectory },
-                { FileDialogAction.PerformAction, PerformAction },
-                { FileDialogAction.Refresh, ForceUpdateItemList },
+                {FileDialogAction.NavigateUp, NavigateUp},
+                {FileDialogAction.NavigateDown, NavigateDown},
+                {FileDialogAction.NavigatePageUp, NavigatePageUp},
+                {FileDialogAction.NavigatePageDown, NavigatePageDown},
+                {FileDialogAction.NavigateToStart, NavigateToStart},
+                {FileDialogAction.NavigateToEnd, NavigateToEnd},
+                {FileDialogAction.NavigateToPreviousDirectory, NavigateToPreviousDirectory},
+                {FileDialogAction.PerformAction, PerformAction},
+                {FileDialogAction.Refresh, ForceUpdateItemList}
             };
         }
 
-        private DrawableItem GetNewDrawableItem(string name, PathItemType type) => new DrawableItem(name, type)
+        private DrawableItem GetNewDrawableItem(string name, PathItemType type)
         {
-            OnSelected = HandleSelection,
-            OnClicked = HandleClick,
-            OnDoubleClicked = HandleDoubleClick,
-        };
+            return new DrawableItem(name, type)
+            {
+                OnSelected = HandleSelection,
+                OnClicked = HandleClick,
+                OnDoubleClicked = HandleDoubleClick
+            };
+        }
     }
 }
