@@ -1,6 +1,8 @@
-﻿using GDE.App.Main.Containers.KeyBindingContainers;
+﻿using System;
+using System.IO;
+using GDAPI.Enumerations;
+using GDE.App.Main.Containers.KeyBindingContainers;
 using GDE.App.Main.Panels;
-using GDAPI.Utilities.Enumerations;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
@@ -8,10 +10,8 @@ using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
 using osuTK;
-using System;
-using System.IO;
 using static GDE.App.Main.Colors.GDEColors;
-using static GDAPI.Utilities.Functions.General.PathExpansionPack;
+using static GDAPI.Functions.General.PathExpansionPack;
 using static System.Environment;
 using static System.Environment.SpecialFolder;
 using static System.IO.Directory;
@@ -22,22 +22,140 @@ namespace GDE.App.Main.UI.FileDialogComponents
     public abstract class FileDialog : Panel
     {
         public const float ItemSpacing = 2.5f;
+        public readonly Bindable<string> CurrentDirectoryBindable = new Bindable<string>("");
 
-        private GDEBreadcrumbNavigationTextBox filePathBreadcrumbs;
-        private TextBox search;
+        public readonly Bindable<DrawableItem> CurrentSelection = new Bindable<DrawableItem>();
 
-        private DirectoryItemContainer itemContainer;
+        private readonly GDEBreadcrumbNavigationTextBox filePathBreadcrumbs;
 
-        protected abstract bool AllowInexistentFileNames { get; }
-        protected virtual string FileDialogActionName { get; set; }
+        private readonly DirectoryItemContainer itemContainer;
+        private readonly TextBox search;
+
+        public readonly Bindable<string> SelectedItemBindable = new Bindable<string>();
 
         /// <summary>The button that performs the file dialog's action.</summary>
         protected GDEButton ActionButton;
 
-        public readonly Bindable<DrawableItem> CurrentSelection = new Bindable<DrawableItem>();
+        public FileDialog(string defaultDirectory = null)
+        {
+            CornerRadius = 10;
+            Masking = true;
 
-        public readonly Bindable<string> SelectedItemBindable = new Bindable<string>();
-        public readonly Bindable<string> CurrentDirectoryBindable = new Bindable<string>("");
+            LockDrag = true;
+
+            AddRangeInternal(new Drawable[]
+            {
+                new Box
+                {
+                    RelativeSizeAxes = Axes.Both,
+                    Colour = FromHex("1a1a1a")
+                },
+                new DrawSizePreservingFillContainer
+                {
+                    RelativeSizeAxes = Axes.Both,
+                    Padding = new MarginPadding(10),
+                    Strategy = DrawSizePreservationStrategy.Minimum,
+                    Children = new Drawable[]
+                    {
+                        new SpriteText
+                        {
+                            Text = $"{FileDialogActionName} File"
+                        },
+                        new Container
+                        {
+                            RelativeSizeAxes = Axes.X,
+                            Y = 35,
+                            Height = 30,
+                            Child = filePathBreadcrumbs = new GDEBreadcrumbNavigationTextBox
+                            {
+                                Origin = Anchor.CentreLeft,
+                                Anchor = Anchor.CentreLeft,
+                                RelativeSizeAxes = Axes.X,
+                                Height = 30
+                            }
+                        },
+                        new FileDialogActionContainer
+                        {
+                            RelativeSizeAxes = Axes.Both,
+                            Padding = new MarginPadding
+                            {
+                                Top = 70,
+                                Bottom = 40
+                            },
+                            Child = itemContainer = new DirectoryItemContainer
+                            {
+                                RelativeSizeAxes = Axes.Both
+                            }
+                        },
+                        new Container
+                        {
+                            Anchor = Anchor.BottomRight,
+                            Origin = Anchor.BottomRight,
+                            RelativeSizeAxes = Axes.X,
+                            Height = 30,
+                            Children = new Drawable[]
+                            {
+                                new Container
+                                {
+                                    Anchor = Anchor.CentreLeft,
+                                    Origin = Anchor.CentreLeft,
+                                    RelativeSizeAxes = Axes.Both,
+                                    Padding = new MarginPadding {Right = 110},
+                                    Children = new Drawable[]
+                                    {
+                                        search = new TextBox
+                                        {
+                                            Anchor = Anchor.CentreLeft,
+                                            Origin = Anchor.CentreLeft,
+                                            RelativeSizeAxes = Axes.Both,
+                                            PlaceholderText = "Selected Path",
+                                            OnCommit = (sender, newText) => SelectedItem = sender.Text
+                                        },
+                                        new SpriteIcon
+                                        {
+                                            Anchor = Anchor.CentreRight,
+                                            Origin = Anchor.CentreRight,
+                                            Size = new Vector2(15),
+                                            Margin = new MarginPadding {Right = 10},
+                                            Icon = FontAwesome.Solid.Search
+                                        }
+                                    }
+                                },
+                                ActionButton = new FadeButton
+                                {
+                                    Anchor = Anchor.CentreRight,
+                                    Origin = Anchor.CentreRight,
+                                    Width = 100,
+                                    EnabledColor = FromHex("303030"),
+                                    Action = ActionButtonAction,
+                                    Text = FileDialogActionName
+                                }
+                            }
+                        }
+                    }
+                }
+            });
+
+            ActionButton.Enabled.Value = false;
+
+            CurrentSelection.BindTo(itemContainer.CurrentSelection);
+            CurrentDirectoryBindable.BindTo(itemContainer.CurrentDirectoryBindable);
+            SelectedItemBindable.BindTo(itemContainer.SelectedItemBindable);
+
+            CurrentSelection.ValueChanged += HandleCurrentSelectionChanged;
+            CurrentDirectoryBindable.ValueChanged += HandleDirectoryChanged;
+            SelectedItemBindable.ValueChanged += HandleItemChanged;
+
+            itemContainer.PerformActionRequested += HandlePerformActionRequested;
+
+            filePathBreadcrumbs.OnTextChanged += HandleBreadcrumbsUpdated;
+            filePathBreadcrumbs.BreadcrumbNavigation.BreadcrumbClicked += HandleBreadcrumbsUpdated;
+
+            CurrentDirectory = defaultDirectory ?? GetFolderPath(MyDocuments);
+        }
+
+        protected abstract bool AllowInexistentFileNames { get; }
+        protected virtual string FileDialogActionName { get; set; }
 
         public bool IsSelectedPathValid
         {
@@ -63,18 +181,21 @@ namespace GDE.App.Main.UI.FileDialogComponents
             }
         }
 
-        public bool CanFinalizeSelection => (AllowInexistentFileNames && IsSelectedPathValid) || CurrentlySelectedItem != null;
+        public bool CanFinalizeSelection =>
+            AllowInexistentFileNames && IsSelectedPathValid || CurrentlySelectedItem != null;
 
         public string SelectedItem
         {
             get => SelectedItemBindable.Value;
             set => SelectedItemBindable.Value = value;
         }
+
         public string SelectedPath
         {
             get => GetCurrentSelectedPath();
             set => UpdatePath(value);
         }
+
         public string CurrentDirectory
         {
             get => CurrentDirectoryBindable.Value;
@@ -86,6 +207,7 @@ namespace GDE.App.Main.UI.FileDialogComponents
             get => CurrentSelection.Value;
             set => CurrentSelection.Value = value;
         }
+
         public int CurrentSelectionIndex
         {
             get => itemContainer.CurrentSelectionIndex;
@@ -94,128 +216,12 @@ namespace GDE.App.Main.UI.FileDialogComponents
 
         public event Action<string> OnFileSelected;
 
-        public FileDialog(string defaultDirectory = null)
-        {
-            CornerRadius = 10;
-            Masking = true;
-
-            LockDrag = true;
-
-            AddRangeInternal(new Drawable[]
-            {
-                new Box
-                {
-                    RelativeSizeAxes = Axes.Both,
-                    Colour = FromHex("1a1a1a")
-                },
-                new DrawSizePreservingFillContainer
-                {
-                    RelativeSizeAxes = Axes.Both,
-                    Padding = new MarginPadding(10),
-                    Strategy = DrawSizePreservationStrategy.Minimum,
-                    Children = new Drawable[]
-                    {
-                        new SpriteText
-                        {
-                            Text = $"{FileDialogActionName} File",
-                        },
-                        new Container
-                        {
-                            RelativeSizeAxes = Axes.X,
-                            Y = 35,
-                            Height = 30,
-                            Child = filePathBreadcrumbs = new GDEBreadcrumbNavigationTextBox
-                            {
-                                Origin = Anchor.CentreLeft,
-                                Anchor = Anchor.CentreLeft,
-                                RelativeSizeAxes = Axes.X,
-                                Height = 30,
-                            }
-                        },
-                        new FileDialogActionContainer
-                        {
-                            RelativeSizeAxes = Axes.Both,
-                            Padding = new MarginPadding
-                            {
-                                Top = 70,
-                                Bottom = 40,
-                            },
-                            Child = itemContainer = new DirectoryItemContainer
-                            {
-                                RelativeSizeAxes = Axes.Both,
-                            },
-                        },
-                        new Container
-                        {
-                            Anchor = Anchor.BottomRight,
-                            Origin = Anchor.BottomRight,
-                            RelativeSizeAxes = Axes.X,
-                            Height = 30,
-                            Children = new Drawable[]
-                            {
-                                new Container
-                                {
-                                    Anchor = Anchor.CentreLeft,
-                                    Origin = Anchor.CentreLeft,
-                                    RelativeSizeAxes = Axes.Both,
-                                    Padding = new MarginPadding { Right = 110 },
-                                    Children = new Drawable[]
-                                    {
-                                        search = new TextBox
-                                        {
-                                            Anchor = Anchor.CentreLeft,
-                                            Origin = Anchor.CentreLeft,
-                                            RelativeSizeAxes = Axes.Both,
-                                            PlaceholderText = "Selected Path",
-                                            OnCommit = (sender, newText) => SelectedItem = sender.Text,
-                                        },
-                                        new SpriteIcon
-                                        {
-                                            Anchor = Anchor.CentreRight,
-                                            Origin = Anchor.CentreRight,
-                                            Size = new Vector2(15),
-                                            Margin = new MarginPadding { Right = 10 },
-                                            Icon = FontAwesome.Solid.Search,
-                                        },
-                                    }
-                                },
-                                ActionButton = new FadeButton
-                                {
-                                    Anchor = Anchor.CentreRight,
-                                    Origin = Anchor.CentreRight,
-                                    Width = 100,
-                                    EnabledColor = FromHex("303030"),
-                                    Action = ActionButtonAction,
-                                    Text = FileDialogActionName,
-                                },
-                            }
-                        }
-                    }
-                }
-            });
-
-            ActionButton.Enabled.Value = false;
-
-            CurrentSelection.BindTo(itemContainer.CurrentSelection);
-            CurrentDirectoryBindable.BindTo(itemContainer.CurrentDirectoryBindable);
-            SelectedItemBindable.BindTo(itemContainer.SelectedItemBindable);
-
-            CurrentSelection.ValueChanged += HandleCurrentSelectionChanged;
-            CurrentDirectoryBindable.ValueChanged += HandleDirectoryChanged;
-            SelectedItemBindable.ValueChanged += HandleItemChanged;
-
-            itemContainer.PerformActionRequested += HandlePerformActionRequested;
-
-            filePathBreadcrumbs.OnTextChanged += HandleBreadcrumbsUpdated;
-            filePathBreadcrumbs.BreadcrumbNavigation.BreadcrumbClicked += HandleBreadcrumbsUpdated;
-
-            CurrentDirectory = defaultDirectory ?? GetFolderPath(MyDocuments);
-        }
-
         public void UpdateActionButtonState()
         {
             ActionButton.Enabled.Value = CanFinalizeSelection;
-            ActionButton.Text = CurrentlySelectedItem?.IsDirectoryOrVolume ?? false ? $"Open {GetUserFriendlyItemName(CurrentlySelectedItem.ItemType)}" : FileDialogActionName;
+            ActionButton.Text = CurrentlySelectedItem?.IsDirectoryOrVolume ?? false
+                ? $"Open {GetUserFriendlyItemName(CurrentlySelectedItem.ItemType)}"
+                : FileDialogActionName;
         }
 
         public void UpdatePath(string newPath)
@@ -230,19 +236,36 @@ namespace GDE.App.Main.UI.FileDialogComponents
             search.Text = value.NewValue;
             UpdateActionButtonState();
         }
+
         public void HandleDirectoryChanged(ValueChangedEvent<string> value)
         {
             UpdateBreadcrumbs();
         }
 
-        protected virtual void ActionButtonAction() => PerformAction();
+        protected virtual void ActionButtonAction()
+        {
+            PerformAction();
+        }
 
-        private void HandleBreadcrumbsUpdated(string dir) => CurrentDirectory = GetCurrentBreadcrumbsDirectory();
+        private void HandleBreadcrumbsUpdated(string dir)
+        {
+            CurrentDirectory = GetCurrentBreadcrumbsDirectory();
+        }
 
-        private void HandleDirectoryChanged() => UpdateBreadcrumbs();
-        private void HandlePerformActionRequested() => PerformAction();
+        private void HandleDirectoryChanged()
+        {
+            UpdateBreadcrumbs();
+        }
 
-        private void HandleCurrentSelectionChanged(ValueChangedEvent<DrawableItem> obj) => UpdateActionButtonState();
+        private void HandlePerformActionRequested()
+        {
+            PerformAction();
+        }
+
+        private void HandleCurrentSelectionChanged(ValueChangedEvent<DrawableItem> obj)
+        {
+            UpdateActionButtonState();
+        }
 
         private void PerformAction()
         {
@@ -252,7 +275,11 @@ namespace GDE.App.Main.UI.FileDialogComponents
                 FinalizeSelection();
         }
 
-        private void NavigateToSelectedDirectory() => CurrentDirectory = GetCurrentSelectedPath();
+        private void NavigateToSelectedDirectory()
+        {
+            CurrentDirectory = GetCurrentSelectedPath();
+        }
+
         private void FinalizeSelection()
         {
             OnFileSelected?.Invoke(SelectedPath);
@@ -266,8 +293,15 @@ namespace GDE.App.Main.UI.FileDialogComponents
             filePathBreadcrumbs.Items.AddRange(dirs);
         }
 
-        private string GetCurrentBreadcrumbsDirectory() => ConcatenateDirectoryPath(filePathBreadcrumbs.Items);
-        private string GetCurrentSelectedPath() => $@"{CurrentDirectory}{CurrentlySelectedItem?.GetPathSuffix() ?? SelectedItem}";
+        private string GetCurrentBreadcrumbsDirectory()
+        {
+            return ConcatenateDirectoryPath(filePathBreadcrumbs.Items);
+        }
+
+        private string GetCurrentSelectedPath()
+        {
+            return $@"{CurrentDirectory}{CurrentlySelectedItem?.GetPathSuffix() ?? SelectedItem}";
+        }
 
         private static string GetUserFriendlyItemName(PathItemType itemType)
         {
@@ -280,6 +314,7 @@ namespace GDE.App.Main.UI.FileDialogComponents
                 case PathItemType.File:
                     return "File";
             }
+
             throw new ArgumentException("Invalid item type.");
         }
     }
